Store line number and char number in token.
This seems like the simplest approach for
getting valuable info in error messages.
It avoids having to keep the source file
in memory (or re-read it) in order to
calculate stuff after an error has occurred.

Also a couple of fixes were I apparently
misunderstood rust, trying to use == to check
token type.